dev-clean-comments
* Clean out commented code
* Replace logger instance
