+++ .gitignore
@@ -21,4 +21,4 @@ package-lock.json
 yarn-error.log
 
 # error reporter
-ErrorLogs
+Error
