--- helper/stackTracer.js
@@ -1,7 +1,0 @@
-import { logger } from './logger';
-
-const stackTracer = (reason) => {
-  logger.error(`${reason.stack}\n`);
-};
-
-export default stackTracer;
