+++ schema/author.js
@@ -1,11 +1,8 @@
 import {
   GraphQLObjectType,
   GraphQLString,
-  // GraphQLList,
   GraphQLInt,
 } from 'graphql';
-// import _ from 'lodash';
-// import BookType from './books';
 
 const AuthorType = new GraphQLObjectType({
   name: 'Author',
@@ -19,12 +16,6 @@ const AuthorType = new GraphQLObjectType({
     age: {
       type: GraphQLInt
     },
-    // books: {
-    //   // type: new GraphQLList(BookType),
-    //   // resolve(parent, args) {
-    //   //   return _.filter(books, { authorId: parent.id})
-    //   // }
-    // },
   })
 });
