+++ utils/authStatusCheck.js
@@ -1,0 +1,7 @@
+const authStatusCheck = (authStatus) => {
+  if (!authStatus) {
+    throw new Error('Permission denied, you need to signup/login');
+  }
+};
+
+export default authStatusCheck;
