+++ utils/initLogger.js
@@ -1,0 +1,4 @@
+import { createLogger } from 'info-logger';
+
+const logger = createLogger('Error', 'log-result');
+export default logger;
